+++ workers/ohsome_quality_analyst/reports/jrc_requirements/report.py
@@ -1,4 +1,6 @@
-from geojson import FeatureCollection
+from typing import Optional, Union
+
+from geojson import MultiPolygon, Polygon
 
 from ohsome_quality_analyst.base.report import BaseReport, IndicatorLayer
 
@@ -6,10 +8,10 @@ from ohsome_quality_analyst.base.report import BaseReport, IndicatorLayer
 class JrcRequirements(BaseReport):
     def __init__(
         self,
-        bpolys: FeatureCollection = None,
-        dataset: str = None,
-        feature_id: int = None,
-        fid_field: str = None,
+        bpolys: Union[Polygon, MultiPolygon, None] = None,
+        dataset: Optional[str] = None,
+        feature_id: Optional[int] = None,
+        fid_field: Optional[str] = None,
     ) -> None:
         """Create a list of indicator objects."""