+++ workers/tests/unittests/test_get_default_figure.py
@@ -13,7 +13,7 @@ class TestGetDefaultFigure(unittest.TestCase):
         infile = os.path.join(
             os.path.dirname(os.path.abspath(__file__)),
             "fixtures",
-            "heidelberg_altstadt.geojson",
+            "heidelberg-altstadt-geometry.geojson",
         )
         with open(infile, "r") as f:
             bpolys = geojson.load(f)